backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-genai>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
